requirement.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1.2
django-allauth==0.51.0
django-cors-headers==3.13.0
django-registration==3.3
django-rest-auth==0.9.5
django-templated-mail==1.1.1
django-webpack-loader==1.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
future==0.18.2
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
MetaTrader5==5.0.36
mpl-finance==0.10.1
oauthlib==3.2.1
openpyxl==3.0.9
pandas-ta==0.3.14b0
Pillow==9.2.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.5.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
stripe==4.2.0
TA-Lib @ file:///C:/Users/GREAT%20SPEED/Desktop/PROJECTS/Part%202/trade%20crypto/TA_Lib-0.4.24-cp39-cp39-win_amd64.whl
types-cryptography==3.3.23
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
